+++ server/endpoints/system.js
@@ -50,7 +50,6 @@ const {
   resetPassword,
   generateRecoveryCodes,
 } = require("../utils/PasswordRecovery");
-const { AzureLoginProvider } = require("../utils/AzureLoginProviders");
 const { SlashCommandPresets } = require("../models/slashCommandsPresets");
 
 function systemEndpoints(app) {
@@ -259,122 +258,6 @@ function systemEndpoints(app) {
       }
     } catch (e) {
       console.error(e.message, e);
-      response.sendStatus(500).end();
-    }
-  });
-  
-  app.post("/azure-login", async (request, response) => {
-    try {
-      if (!(await SystemSettings.isMultiUserMode())) {
-        response.status(200).json({
-          user: null,
-          valid: false,
-          token: null,
-          message: "Azure login is available on multi-user mode",
-        });
-        return;
-      }
-
-      const data = reqBody(request);
-      const azureLoginProvider = new AzureLoginProvider();
-      const { username } = await azureLoginProvider.login(data);
-      let user = await User.get({ username: String(username) });
-
-      const allowedDomain = (
-        await SystemSettings.get({ label: "allowed_domain" })
-      )?.value;
-      if (allowedDomain && allowedDomain !== username.split("@")[1]) {
-        await EventLogs.logEvent(
-          "failed_login_domain_not_allowed",
-          {
-            ip: request.ip || "Unknown IP",
-            username: username || "Unknown user",
-          },
-          user?.id
-        );
-        response.status(200).json({
-          user: null,
-          valid: false,
-          token: null,
-          message: "[006] Domain not allowed by admin.",
-        });
-        return;
-      }
-
-      if (!user) {
-        const { user: newUser, error } = await User.createWithAzureLoginProvider({
-          username: String(username),
-        });
-        if (!newUser) {
-          await EventLogs.logEvent(
-            "failed_login_error_creating_user",
-            {
-              ip: request.ip || "Unknown IP",
-              username: username || "Unknown user",
-            },
-            existingUser?.id
-          );
-          response.status(200).json({
-            user: null,
-            valid: false,
-            token: null,
-            message: error,
-          });
-          return;
-        }
-        await EventLogs.logEvent(
-          "user_created",
-          {
-            userName: newUser.username,
-            createdBy: newUser.username,
-          },
-          newUser.id
-        );
-        user = newUser;
-      }
-
-      if (user.suspended) {
-        await EventLogs.logEvent(
-          "failed_login_account_suspended",
-          {
-            ip: request.ip || "Unknown IP",
-            username: username || "Unknown user",
-          },
-          user?.id
-        );
-        response.status(200).json({
-          user: null,
-          valid: false,
-          token: null,
-          message: "[004] Account suspended by admin.",
-        });
-        return;
-      }
-
-      await Telemetry.sendTelemetry(
-        "login_event",
-        { multiUserMode: false },
-        user?.id
-      );
-
-      await EventLogs.logEvent(
-        "login_event",
-        {
-          ip: request.ip || "Unknown IP",
-          username: user.username || "Unknown user",
-        },
-        user?.id
-      );
-
-      response.status(200).json({
-        valid: true,
-        user: user,
-        token: makeJWT({ id: user.id, username: user.username }, "30d"),
-        message: null,
-      });
-      return;
-    } catch (e) {
-      console.log(e.message, e);
       response.sendStatus(500).end();
     }
   });
